% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censlm-package.R
\docType{package}
\name{censlm-package}
\alias{censlm}
\alias{censlm-package}
\title{censlm: Censored Linear Models}
\description{
Fit linear models in the presence of censored data. Simulate plausible values for censored observations based on a fitted model.
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@protonmail.com}

}
\keyword{internal}
