% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clm.R
\name{clm}
\alias{clm}
\title{Fit censored linear models}
\usage{
clm(formula, data, ..., left)
}
\description{
A specialized interface to \code{\link[survival:survreg]{survival::survreg()}}, \code{clm()} fits linear models
in the presence of censored observations.
}
